pyproject.toml
[project]
name = "gravitic-aphelion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ddgs", # Renamed from duckduckgo-search>=8.1.1
    "fastapi>=0.122.0",
    "langchain>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.2.0",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.0",
    "plotly>=6.5.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
    "yfinance>=0.2.66",
    # Added dependencies for technical analysis tools
    "pandas>=2.2.2", 
    "numpy>=1.26.4", 
]

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.4.7",
    "pytest>=9.0.1",
]